change logging over to system events and remove monolog.improve crud.

+++ Src/Infrastructure/Database/CRUD/CrudController.php
@@ -6,7 +6,6 @@ namespace It_All\Spaghettify\Src\Infrastructure\Database\CRUD;
 use It_All\Spaghettify\Src\Infrastructure\Controller;
 use It_All\Spaghettify\Src\Infrastructure\UserInterface\Forms\FormHelper;
 use function It_All\Spaghettify\Src\Infrastructure\Utilities\getRouteName;
-use It_All\Spaghettify\Src\Infrastructure\Utilities\SimpleValidatorExtension;
 use Slim\Container;
 use Slim\Http\Request;
 use Slim\Http\Response;
@@ -57,8 +56,10 @@ class CrudController extends Controller
             return $this->view->insertView($request, $response, $args);
         }
 
-        if (!$this->insert()) {
-            throw new \Exception("Insert Failure");
+        try {
+            $this->insert();
+        } catch (\Exception $e) {
+            throw new \Exception("Insert failure. ".$e->getMessage(), E_USER_ERROR);
         }
 
         FormHelper::unsetSessionVars();
@@ -87,6 +88,8 @@ class CrudController extends Controller
 
         // make sure there is a record for the primary key in the model
         if (!$record = $this->model->selectForPrimaryKey($args['primaryKey'])) {
+            $eventNote = $this->model->getPrimaryKeyColumnName().": ".$args['primaryKey'].", Table: ".$this->model->getTableName();
+            $this->systemEvents->insertWarning('Record not found for update', (int) $this->authentication->getUserId(), $eventNote);
             $_SESSION[SESSION_ADMIN_NOTICE] = ["Record ".$args['primaryKey']." Not Found", 'adminNoticeFailure'];
             return $response->withRedirect($this->router->pathFor($redirectRoute));
         }
@@ -125,8 +128,10 @@ class CrudController extends Controller
             return $this->view->updateView($request, $response, $args);
         }
 
-        if (!$this->update($response, $args)) {
-            throw new \Exception("Update Failure");
+        try {
+            $this->update($response, $args);
+        } catch (\Exception $e) {
+            throw new \Exception("Update failure. ".$e->getMessage(), E_USER_ERROR);
         }
 
         FormHelper::unsetSessionVars();
@@ -135,7 +140,18 @@ class CrudController extends Controller
 
     public function getDelete(Request $request, Response $response, $args)
     {
-        return $this->delete($response, $args);
+        if (!$this->authorization->checkFunctionality(getRouteName(true, $this->routePrefix, 'delete'))) {
+            throw new \Exception('No permission.');
+        }
+
+        try {
+            $this->delete($response, $args);
+        } catch (\Exception $e) {
+            // no need to do anything, just redirect with error message already set
+        }
+
+        $redirectRoute = getRouteName(true, $this->routePrefix, 'index');
+        return $response->withRedirect($this->router->pathFor($redirectRoute));
     }
 
     /**
@@ -161,18 +177,23 @@ class CrudController extends Controller
         try {
             $res = $this->model->insertRecord($_SESSION[SESSION_REQUEST_INPUT_KEY]);
             $returned = pg_fetch_all($res);
-            $message = 'Inserted record '.$returned[0][$this->model->getPrimaryKeyColumnName()].
-                ' into '.$this->model->getTableName();
-            $this->logger->addInfo($message);
+            $primaryKeyColumnName = $this->model->getPrimaryKeyColumnName();
+            $insertedRecordId = $returned[0][$primaryKeyColumnName];
+            $tableName = $this->model->getTableName();
+
+            $eventNote = "$primaryKeyColumnName: $insertedRecordId, Table: $tableName";
+            $this->systemEvents->insertInfo('Inserted database record', (int) $this->authentication->getUserId(), $eventNote);
+
             if ($sendEmail) {
                 $settings = $this->container->get('settings');
                 $this->mailer->send(
                     $_SERVER['SERVER_NAME'] . " Event",
-                    "Inserted into ".$this->model->getTableName()."\n See event log for details.",
-                    [$settings['emails']['owner']]
+                    "Inserted into $tableName.\n See event log for details.",
+                    [$settings['emails']['programmer']]
                 );
             }
-            $_SESSION[SESSION_ADMIN_NOTICE] = [$message, 'adminNoticeSuccess'];
+
+            $_SESSION[SESSION_ADMIN_NOTICE] = ["Inserted record $insertedRecordId", 'adminNoticeSuccess'];
 
             return true;
 
@@ -181,14 +202,29 @@ class CrudController extends Controller
         }
     }
 
-    protected function update(Response $response, $args)
+    protected function update(Response $response, $args, bool $sendEmail = false)
     {
         // attempt to update the model
         try {
             $this->model->updateRecordByPrimaryKey($_SESSION[SESSION_REQUEST_INPUT_KEY], $args['primaryKey']);
-            $message = 'Updated record '.$args['primaryKey'];
-            $this->logger->addInfo($message . ' in '. $this->model->getTableName());
-            $_SESSION[SESSION_ADMIN_NOTICE] = [$message, 'adminNoticeSuccess'];
+
+            $primaryKeyColumnName = $this->model->getPrimaryKeyColumnName();
+            $updatedRecordId = $args['primaryKey'];
+            $tableName = $this->model->getTableName();
+
+            $eventNote = "$primaryKeyColumnName: $updatedRecordId, Table: $tableName";
+            $this->systemEvents->insertInfo('Updated database record', (int) $this->authentication->getUserId(), $eventNote);
+
+            if ($sendEmail) {
+                $settings = $this->container->get('settings');
+                $this->mailer->send(
+                    $_SERVER['SERVER_NAME'] . " Event",
+                    "Updated $tableName.\n See event log for details.",
+                    [$settings['emails']['programmer']]
+                );
+            }
+
+            $_SESSION[SESSION_ADMIN_NOTICE] = ["Updated record $updatedRecordId", 'adminNoticeSuccess'];
 
             return true;
 
@@ -199,42 +235,36 @@ class CrudController extends Controller
 
     protected function delete(Response $response, $args, string $returnColumn = null, bool $sendEmail = false)
     {
-        if (!$this->authorization->checkFunctionality(getRouteName(true, $this->routePrefix, 'delete'))) {
-            throw new \Exception('No permission.');
-        }
-
         $primaryKey = $args['primaryKey'];
-        $redirectRoute = getRouteName(true, $this->routePrefix, 'index');
+        $primaryKeyColumnName = $this->model->getPrimaryKeyColumnName();
+        $tableName = $this->model->getTableName();
+        $eventNote = "$primaryKeyColumnName: $primaryKey, Table: $tableName";
 
-        if ($res = $this->model->deleteByPrimaryKey($primaryKey, $returnColumn)) {
-            $message = 'Deleted record '.$primaryKey;
-            if ($returnColumn != null) {
-                $returned = pg_fetch_all($res);
-                $message .= ' ('.$returnColumn.' '.$returned[0][$returnColumn].')';
-            }
-            $this->logger->addInfo($message . " from ".$this->model->getTableName()."");
-            if ($sendEmail) {
-                $settings = $this->container->get('settings');
-                $this->mailer->send(
-                    $_SERVER['SERVER_NAME'] . " Event",
-                    "Deleted record from ".$this->model->getTableName().".\nSee event log for details.",
-                    [$settings['emails']['owner']]
-                );
-            }
-            $_SESSION[SESSION_ADMIN_NOTICE] = [$message, 'adminNoticeSuccess'];
-
-            return $response->withRedirect($this->router->pathFor($redirectRoute));
-
-        } else {
-
-            $this->logger->addWarning("primary key $primaryKey for ".$this->model->getTableName()." not found for deletion. IP: " . $_SERVER['REMOTE_ADDR']);
-
-            $settings = $this->container->get('settings');
-            $this->mailer->send($_SERVER['SERVER_NAME'] . " Event", "primary key $primaryKey not found for deletion. Check event log for details.", [$settings['emails']['programmer']]);
-
+        if (!$res = $this->model->deleteByPrimaryKey($primaryKey, $returnColumn)) {
+            $this->systemEvents->insertWarning('Primary key not found for delete', (int) $this->authentication->getUserId(), $eventNote);
             $_SESSION[SESSION_ADMIN_NOTICE] = [$primaryKey.' not found', 'adminNoticeFailure'];
-
-            return $response->withRedirect($this->router->pathFor($redirectRoute));
+            return false;
         }
+
+        $adminMessage = 'Deleted record '.$primaryKey;
+        if ($returnColumn != null) {
+            $returned = pg_fetch_all($res);
+            $eventNote .= ", $returnColumn: ".$returned[0][$returnColumn];
+            $adminMessage .= " ($returnColumn ".$returned[0][$returnColumn].")";
+        }
+
+        $this->systemEvents->insertInfo('Deleted database record', (int) $this->authentication->getUserId(), $eventNote);
+
+        if ($sendEmail) {
+            $settings = $this->container->get('settings');
+            $this->mailer->send(
+                $_SERVER['SERVER_NAME'] . " Event",
+                "Deleted record from $tableName.\nSee event log for details.",
+                [$settings['emails']['programmer']]
+            );
+        }
+
+        $_SESSION[SESSION_ADMIN_NOTICE] = [$adminMessage, 'adminNoticeSuccess'];
+        return true;
     }
 }
